Adding tests for ALU Opcode enums
