+++ src/test/java/org/karnes/homebrew/emulator/component/alu/LogicUnitTest.java
@@ -1,4 +1,4 @@
-package org.karnes.homebrew.emulator.component;
+package org.karnes.homebrew.emulator.component.alu;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -6,20 +6,20 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 import org.karnes.homebrew.bitset.FixedBitSet;
 import org.karnes.homebrew.emulator.ConditionCode;
-import org.karnes.homebrew.emulator.component.alu.arithmeticunit.AU_OPCODE;
-import org.karnes.homebrew.emulator.component.alu.arithmeticunit.ArithmeticUnit;
 import org.karnes.homebrew.emulator.component.bus.BidirectionalBus;
 import org.karnes.homebrew.emulator.component.bus.VirtualBus;
 import org.karnes.homebrew.emulator.component.bus.connection.ReadableBusConnection;
 import org.karnes.homebrew.emulator.component.bus.connection.WriteableBusConnection;
+import org.karnes.homebrew.emulator.component.alu.logicunit.LU_OPCODE;
+import org.karnes.homebrew.emulator.component.alu.logicunit.LogicUnit;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ArithemeticUnitTest {
+class LogicUnitTest {
     private static int DATA_WIDTH = 4;
 
-    private ArithmeticUnit arithmeticUnit;
-    private WriteableBusConnection auOperationBusConnection;
+    private LogicUnit logicUnit;
+    private WriteableBusConnection luOperationBusConnection;
     private WriteableBusConnection tmp1BusConnection;
     private WriteableBusConnection tmp2BusConnection;
     private ReadableBusConnection outputBusConnection;
@@ -27,26 +27,26 @@ class ArithemeticUnitTest {
 
     @BeforeEach
     void setUp() {
-        BidirectionalBus auOperationBus = new VirtualBus("AU_OPERATION_BUS", AU_OPCODE.WIDTH);
-        auOperationBusConnection = auOperationBus.getWriteConnection();
+        BidirectionalBus luOperationBus = new VirtualBus("LU_OPERATION_BUS", LU_OPCODE.WIDTH);
+        luOperationBusConnection = luOperationBus.getWriteConnection();
         BidirectionalBus tmp1Bus = new VirtualBus("TMP1-ALU_BUS", DATA_WIDTH);
         tmp1BusConnection = tmp1Bus.getWriteConnection();
         BidirectionalBus tmp2Bus = new VirtualBus("TMP2-ALU_BUS", DATA_WIDTH);
         tmp2BusConnection = tmp2Bus.getWriteConnection();
-        BidirectionalBus outputBus = new VirtualBus("AU_OUTPUT_BUS", DATA_WIDTH);
+        BidirectionalBus outputBus = new VirtualBus("LU_OUTPUT_BUS", DATA_WIDTH);
         outputBusConnection = outputBus.getReadConnection();
         BidirectionalBus ccBus = new VirtualBus("CC_BUS", ConditionCode.WIDTH);
         ccBusConnection = ccBus.getReadConnection();
 
-        arithmeticUnit = new ArithmeticUnit(auOperationBus, tmp1Bus, tmp2Bus, outputBus, ccBus);
+        logicUnit = new LogicUnit(luOperationBus, tmp1Bus, tmp2Bus, outputBus, ccBus);
     }
 
 
     @Test
     void simpleTest() {
-        //The AU is already hooked up.
-        //Enable the output. This should be AU Enable + ADD (0,0)
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.ADD.toBitSet());
+        //The LU is already hooked up.
+        //Enable the output. This should be LU Enable + XOR (0,0)
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.XOR.toBitSet());
 
         //The output should be 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
@@ -64,52 +64,52 @@ class ArithemeticUnitTest {
 
     @Test
     void testChangingOperations() {
-        //The AU is already hooked up.
+        //The LU is already hooked up.
         //First send the OFF signal
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.OFF.toBitSet());
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OFF.toBitSet());
 
         //The output and cc should be 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet(DATA_WIDTH), ccBusConnection.readBusValue());
 
         //Set some inputs
-        tmp1BusConnection.writeValueToBus(new FixedBitSet("0101")); // 5
-        tmp2BusConnection.writeValueToBus(new FixedBitSet("1010")); // -6
+        tmp1BusConnection.writeValueToBus(new FixedBitSet("0101"));
+        tmp2BusConnection.writeValueToBus(new FixedBitSet("1010"));
 
         //The output and cc should be still 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet(DATA_WIDTH), ccBusConnection.readBusValue());
 
-        //Enable ADD
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.ADD.toBitSet());
+        //Enable OR
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OR.toBitSet());
 
         //The output and cc should now have valid values.
-        assertEquals(FixedBitSet.allOnes(DATA_WIDTH), outputBusConnection.readBusValue()); //-1
+        assertEquals(FixedBitSet.allOnes(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet("0010"), ccBusConnection.readBusValue());
 
-        //Change to SUB
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.SUB.toBitSet());
+        //Change to AND
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.AND.toBitSet());
 
         //The output and cc should change accordingly.
-        assertEquals(new FixedBitSet("1011"), outputBusConnection.readBusValue()); //is 11 -> overflow to: -5
-        assertEquals(new FixedBitSet("1110"), ccBusConnection.readBusValue());
+        assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
+        assertEquals(new FixedBitSet("0001"), ccBusConnection.readBusValue());
 
-        //Change to INC
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.INC.toBitSet());
-
-        //The output and cc should now have valid values.
-        assertEquals(new FixedBitSet("0110"), outputBusConnection.readBusValue()); //6
-        assertEquals(new FixedBitSet("0000"), ccBusConnection.readBusValue());
-
-        //Change to DEC
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.DEC.toBitSet());
+        //Change to XOR
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.XOR.toBitSet());
 
         //The output and cc should now have valid values.
-        assertEquals(new FixedBitSet("0100"), outputBusConnection.readBusValue()); //4
-        assertEquals(new FixedBitSet("0000"), ccBusConnection.readBusValue());
+        assertEquals(FixedBitSet.allOnes(DATA_WIDTH), outputBusConnection.readBusValue());
+        assertEquals(new FixedBitSet("0010"), ccBusConnection.readBusValue());
+
+        //Change to NOT
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.NOT.toBitSet());
+
+        //The output and cc should now have valid values.
+        assertEquals(new FixedBitSet("1010"), outputBusConnection.readBusValue());
+        assertEquals(new FixedBitSet("0010"), ccBusConnection.readBusValue());
 
         //Turn it back to OFF
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.OFF.toBitSet());
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OFF.toBitSet());
 
         //The output and cc should be 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
@@ -118,39 +118,39 @@ class ArithemeticUnitTest {
 
     @Test
     void testChangingInputs() {
-        //The AU is already hooked up.
+        //The LU is already hooked up.
         //First send the OFF signal
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.OFF.toBitSet());
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OFF.toBitSet());
 
         //The output and cc should be 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet(DATA_WIDTH), ccBusConnection.readBusValue());
 
         //Set some inputs
-        tmp1BusConnection.writeValueToBus(new FixedBitSet("0101")); // 5
-        tmp2BusConnection.writeValueToBus(new FixedBitSet("1010")); // -6
+        tmp1BusConnection.writeValueToBus(new FixedBitSet("0101"));
+        tmp2BusConnection.writeValueToBus(new FixedBitSet("1010"));
 
         //The output and cc should be still 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet(DATA_WIDTH), ccBusConnection.readBusValue());
- 
-        //Enable ADD
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.ADD.toBitSet());
+
+        //Enable OR
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OR.toBitSet());
 
         //The output and cc should now have valid values.
-        assertEquals(FixedBitSet.allOnes(DATA_WIDTH), outputBusConnection.readBusValue()); //-1
+        assertEquals(FixedBitSet.allOnes(DATA_WIDTH), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet("0010"), ccBusConnection.readBusValue());
 
         //Change the first input
-        tmp1BusConnection.writeValueToBus(new FixedBitSet("0000")); // 0
+        tmp1BusConnection.writeValueToBus(new FixedBitSet("0000"));
 
         //The output and cc should change accordingly.
-        assertEquals(new FixedBitSet("1010"), outputBusConnection.readBusValue()); // -6
+        assertEquals(new FixedBitSet("1010"), outputBusConnection.readBusValue());
         assertEquals(new FixedBitSet("0010"), ccBusConnection.readBusValue());
 
 
         //Change the second input
-        tmp2BusConnection.writeValueToBus(new FixedBitSet("0000")); //0
+        tmp2BusConnection.writeValueToBus(new FixedBitSet("0000"));
 
         //The output and cc should change accordingly.
         assertEquals(new FixedBitSet("0000"), outputBusConnection.readBusValue());
@@ -165,7 +165,7 @@ class ArithemeticUnitTest {
 
 
         //Turn it back to OFF
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.OFF.toBitSet());
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OFF.toBitSet());
 
         //The output and cc should be 0.
         assertEquals(new FixedBitSet(DATA_WIDTH), outputBusConnection.readBusValue());
@@ -174,15 +174,15 @@ class ArithemeticUnitTest {
 
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/4bit_ADDs.csv", numLinesToSkip = 1)
-    void testAllADDs(FixedBitSet a, FixedBitSet b, FixedBitSet result, FixedBitSet cc) {
-        //The AU is already hooked up.
+    @CsvFileSource(resources = "/4bit_XORs.csv", numLinesToSkip = 1)
+    void testAllXORs(FixedBitSet a, FixedBitSet b, FixedBitSet result, FixedBitSet cc) {
+        //The LU is already hooked up.
         //Set the inputs
         tmp1BusConnection.writeValueToBus(a);
         tmp2BusConnection.writeValueToBus(b);
 
-        //Enable the output. This should be AU Enable + ADD (a,b)
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.ADD.toBitSet());
+        //Enable the output. This should be LU Enable + XOR (a,b)
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.XOR.toBitSet());
 
         //The output should be result.
         assertEquals(result, outputBusConnection.readBusValue());
@@ -194,15 +194,15 @@ class ArithemeticUnitTest {
 
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/4bit_SUBs.csv", numLinesToSkip = 1)
-    void testAllSUBs(FixedBitSet a, FixedBitSet b, FixedBitSet result, FixedBitSet cc) {
-        //The AU is already hooked up.
+    @CsvFileSource(resources = "/4bit_ORs.csv", numLinesToSkip = 1)
+    void testAllORs(FixedBitSet a, FixedBitSet b, FixedBitSet result, FixedBitSet cc) {
+        //The LU is already hooked up.
         //Set the inputs
         tmp1BusConnection.writeValueToBus(a);
         tmp2BusConnection.writeValueToBus(b);
 
-        //Enable the output. This should be AU Enable + SUB (a,b)
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.SUB.toBitSet());
+        //Enable the output. This should be LU Enable + OR (a,b)
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.OR.toBitSet());
 
         //The output should be result.
         assertEquals(result, outputBusConnection.readBusValue());
@@ -213,14 +213,15 @@ class ArithemeticUnitTest {
     }
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/4bit_INCs.csv", numLinesToSkip = 1)
-    void testAllINCs(FixedBitSet a, FixedBitSet result, FixedBitSet cc) {
-        //The AU is already hooked up.
+    @CsvFileSource(resources = "/4bit_ANDs.csv", numLinesToSkip = 1)
+    void testAllANDs(FixedBitSet a, FixedBitSet b, FixedBitSet result, FixedBitSet cc) {
+        //The LU is already hooked up.
         //Set the inputs
         tmp1BusConnection.writeValueToBus(a);
+        tmp2BusConnection.writeValueToBus(b);
 
-        //Enable the output. This should be AU Enable + INC (a,b)
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.INC.toBitSet());
+        //Enable the output. This should be LU Enable + AND (a,b)
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.AND.toBitSet());
 
         //The output should be result.
         assertEquals(result, outputBusConnection.readBusValue());
@@ -231,14 +232,14 @@ class ArithemeticUnitTest {
     }
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/4bit_DECs.csv", numLinesToSkip = 1)
-    void testAllDECs(FixedBitSet a, FixedBitSet result, FixedBitSet cc) {
-        //The AU is already hooked up.
+    @CsvFileSource(resources = "/4bit_NOTs.csv", numLinesToSkip = 1)
+    void testAllNOTs(FixedBitSet a, FixedBitSet result, FixedBitSet cc) {
+        //The LU is already hooked up.
         //Set the inputs
         tmp1BusConnection.writeValueToBus(a);
 
-        //Enable the output. This should be AU Enable + DEC (a)
-        auOperationBusConnection.writeValueToBus(AU_OPCODE.DEC.toBitSet());
+        //Enable the output. This should be LU Enable + NOT (a)
+        luOperationBusConnection.writeValueToBus(LU_OPCODE.NOT.toBitSet());
 
         //The output should be result.
         assertEquals(result, outputBusConnection.readBusValue());
